# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Video Streaming Service API
  version: 1.0.0
  contact:
    name: video-streaming-service
    url: https://github.com/Nuu-maan/video-streaming-service
  description: |
    Standalone JSON API for a video platform: accounts, uploads, transcoding,
    HLS/MP4 playback, comments, playlists, subscriptions, search, moderation
    and admin analytics. It is designed to be hosted on its own and consumed
    by a separate frontend (React, Next.js, Vue, ...) running on a different
    origin.

    ## The three things to know first

    1. **Auth is JWT Bearer.** Log in at `POST /auth/login` (the body field is
       `identifier` — a username *or* an email — plus `password`) and send
       `Authorization: Bearer <access_token>` on every authenticated call.
       Access tokens live 900 s; exchange the `refresh_token` at
       `POST /auth/refresh` for a new pair. A refresh token is *not* usable as
       an API credential, and an access token is rejected at `/auth/refresh`.

    2. **Every JSON response is wrapped in an envelope.**
       Success: `{"success": true, "data": ...}`.
       Lists (including `/search`): `{"success": true, "data": [...], "pagination": {"total", "page", "limit", "total_pages", "has_next", "has_previous"}}`.
       Errors: `{"success": false, "error": {"code": "...", "message": "..."}}`.
       Media routes (`.m3u8`, `.ts`, MP4, JPEG) return raw bytes, never the envelope.

...

    ## Other conventions

...
servers:
...
    variables:
...
security: []
paths:
...
  /admin/videos/{id}/cache:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    delete:
      tags: [Admin]
      operationId: clearPlaylistCache
      summary: Flush the cached HLS playlists for a video
      description: Requires `moderate_content`.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cache cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <access_token>`. Obtain a token pair from
        `/auth/login` or `/auth/register`; renew with `/auth/refresh`.
        Access tokens live 900 s. A refresh token is not accepted here.
  parameters:
...
    VideoId:
      name: id
      in: path
      required: true
      description: Video id
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Malformed input (`VALIDATION_ERROR`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid request
    Unauthorized:
      description: Missing, invalid, expired or revoked access token (`UNAUTHORIZED`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
    Forbidden:
      description: Authenticated, but lacking the required permission (`FORBIDDEN`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Insufficient permissions
...
  schemas:
...
    SuccessEnvelope:
      type: object
      description: 'Every non-media, non-list success response: `{"success": true, "data": ...}`'
      required: [success]
      properties:
        success:
          type: boolean
          const: true
        data:
          description: The payload; its shape is documented per endpoint
...
    ErrorResponse:
      type: object
      description: 'Every error: `{"success": false, "error": {"code", "message"}}`'
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: "#/components/schemas/ErrorDetail"
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >-
            Machine-readable error code. `AUTH_UNAVAILABLE` arrives with
            status 503 when the token-revocation store is unreachable and the
            server fails closed — retry with the same token rather than
            discarding credentials as you would on a 401.
          enum:
            - VALIDATION_ERROR
            - BAD_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - INTERNAL_ERROR
            - ALREADY_EXISTS
            - USER_BANNED
            - DUPLICATE_REPORT
            - EMAIL_ALREADY_VERIFIED
            - INVALID_TOKEN
            - INVALID_CURRENT_PASSWORD
            - FILE_TOO_LARGE
            - INVALID_FORMAT
            - HLS_NOT_READY
            - PLAYLIST_NOT_FOUND
            - SEGMENT_NOT_FOUND
            - VIDEO_NOT_READY
            - FILE_NOT_FOUND
            - ALREADY_IN_PLAYLIST
            - AUTH_UNAVAILABLE
        message:
          type: string
          description: Human-readable explanation
    MessageResponse:
      description: A success envelope whose payload is a short confirmation
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                message:
                  type: string
              additionalProperties: true

    # ── Enums ──
...